Refactor layout to include ThemeProvider and Navbar components; remove theme management from Home component for cleaner structure.
