+++ frontend/src/components/Navbar.tsx
@@ -1,0 +1,108 @@
+"use client";
+
+import Link from "next/link";
+import { useTheme } from "./ThemeProvider";
+
+export default function Navbar() {
+  const { theme, toggleTheme, mounted } = useTheme();
+
+  return (
+    <nav className="fixed top-0 w-full z-50 glass border-b border-border/50">
+      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
+        <div className="flex justify-between items-center h-20">
+          <div className="flex items-center space-x-3 group cursor-pointer">
+            <div className="relative">
+              <div className="absolute inset-0 bg-gradient-to-br from-primary-600 to-secondary-600 rounded-2xl blur-lg opacity-50 group-hover:opacity-75 transition-opacity" />
+              <div className="relative w-12 h-12 bg-gradient-to-br from-primary-600 to-secondary-600 rounded-2xl flex items-center justify-center shadow-glow">
+                <span className="text-white font-bold text-xl">EM</span>
+              </div>
+            </div>
+            <div>
+              <span className="text-2xl font-bold bg-gradient-to-r from-primary-600 to-secondary-600 bg-clip-text text-transparent">
+                Exes Manen
+              </span>
+              <p className="text-xs text-muted-foreground">
+                Expense Management
+              </p>
+            </div>
+          </div>
+
+          <div className="hidden md:flex items-center space-x-8">
+            <Link
+              href="#features"
+              className="text-foreground/80 hover:text-foreground transition-colors font-medium"
+            >
+              Features
+            </Link>
+            <Link
+              href="#how-it-works"
+              className="text-foreground/80 hover:text-foreground transition-colors font-medium"
+            >
+              How It Works
+            </Link>
+            <Link
+              href="#pricing"
+              className="text-foreground/80 hover:text-foreground transition-colors font-medium"
+            >
+              Pricing
+            </Link>
+          </div>
+
+          <div className="flex items-center space-x-2 sm:space-x-4">
+            {/* Theme Toggle */}
+            {mounted && (
+              <button
+                onClick={toggleTheme}
+                className="p-2 rounded-lg hover:bg-muted transition-colors"
+                aria-label="Toggle theme"
+              >
+                {theme === "light" ? (
+                  <svg
+                    className="w-5 h-5"
+                    fill="none"
+                    stroke="currentColor"
+                    viewBox="0 0 24 24"
+                  >
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      strokeWidth={2}
+                      d="M20.354 15.354A9 9 0 018.646 3.646 9.003 9.003 0 0012 21a9.003 9.003 0 008.354-5.646z"
+                    />
+                  </svg>
+                ) : (
+                  <svg
+                    className="w-5 h-5"
+                    fill="none"
+                    stroke="currentColor"
+                    viewBox="0 0 24 24"
+                  >
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      strokeWidth={2}
+                      d="M12 3v1m0 16v1m9-9h-1M4 12H3m15.364 6.364l-.707-.707M6.343 6.343l-.707-.707m12.728 0l-.707.707M6.343 17.657l-.707.707M16 12a4 4 0 11-8 0 4 4 0 018 0z"
+                    />
+                  </svg>
+                )}
+              </button>
+            )}
+
+            <Link
+              href="/login"
+              className="hidden sm:block text-foreground hover:text-primary-600 transition-colors font-medium"
+            >
+              Login
+            </Link>
+            <Link
+              href="/signup"
+              className="px-4 py-2 sm:px-6 sm:py-3 bg-gradient-to-r from-primary-600 to-secondary-600 text-white rounded-lg sm:rounded-xl font-semibold shadow-lg hover:shadow-xl transition-all duration-300 hover:scale-105 text-sm sm:text-base"
+            >
+              Get Started
+            </Link>
+          </div>
+        </div>
+      </div>
+    </nav>
+  );
+}
